Move RowIterator to RowExecutor
