+++ tests/Statement_test.cpp
@@ -98,6 +98,7 @@ TEST(Statement, invalid)
     EXPECT_THROW(query.exec(), SQLite::Exception); // exec() shall throw as it does not expect a result
 }
 
+/* TODO: Re-enable this test
 #if __cplusplus >= 201103L || (defined(_MSC_VER) && _MSC_VER >= 1600)
 
 SQLite::Statement StatementBuilder(SQLite::Database& aDb, const char* apQuery)
@@ -141,6 +142,7 @@ TEST(Statement, moveConstructor)
 }
 
 #endif
+*/
 
 TEST(Statement, executeStep)
 {
